update 'require' paths to use 'specs' folder instead of 'tests'
I had changed the name of this directory at the last minute for organizational consistency, and in my haste I neglected to update the paths that require spec_helpers.rb from within the spec files themselves.
This commit correctly requires 'spec_helpers' from '.specs/spec_helpers' so the tests are able to run.